requirements.txt
attr==0.3.2
azureml_core==1.56.0
azureml_dataprep_native==41.0.0
azureml_dataprep_rslex==2.22.2
azureml_telemetry==1.56.0
azureml_widgets==1.56.0
colorama==0.4.6
ConfigParser==7.0.0
cryptography==43.0.0
Cython==3.0.10
dl==0.1.0
docutils==0.21.2
filelock==3.15.4
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==7.8.2
jnius==1.1.0
joblib==1.4.2
keyring==25.2.1
matplotlib==3.9.1
numpy==2.0.1
pandas==2.2.2
Pillow==10.4.0
protobuf==5.27.2
pyOpenSSL==24.2.1
pyOpenSSL==24.2.1
redis==5.0.7
scikit_learn==1.5.1
Sphinx==7.4.7
thread==2.0.5
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
trove_classifiers==2024.7.2
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
